# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flock Energy - Urja Meter Ops API",
    "description": "\nA clean, documented REST API over the legacy **Urja Meter Ops** portal.\n\nThe portal has no API. This service logs in as a normal user, manages the session,\nreverse-engineers the portal's internal endpoints, normalises its inconsistent payloads,\nand exposes the result as typed JSON that another program can consume without ever\ntouching the portal.\n\n**What it adds over the portal**\n\n* Filtering, sorting and pagination the portal cannot do (make, status, phase, build,\n  any hierarchy code).\n* The network hierarchy, reconstructed from per-meter paths into a queryable tree.\n* Real consumption, derived by differencing the portal's cumulative registers.\n* Proximity search over meter coordinates.\n* An explicit data-quality report instead of silently tidied-away inconsistencies.\n\nSee `PROTOCOL.md` for how the portal actually works, and `README.md` for design decisions\nand trade-offs.\n",
    "contact": {
      "name": "Flock Energy take-home submission"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/meters": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "List meters",
        "description": "Filter, sort and page the full meter estate.\n\nServed from the local snapshot, so this supports filters the portal itself cannot: by make, status, phase, install type, build and any hierarchy code. All filters are ANDed. `search` matches meter id or serial as a literal substring - unlike the portal, `%` and `_` are not wildcards here.",
        "operationId": "list_meters_api_v1_meters_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Literal substring of meter id or serial.",
              "title": "Search"
            },
            "description": "Literal substring of meter id or serial."
          },
          {
            "name": "make",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact manufacturer, e.g. 'HPL'.",
              "title": "Make"
            },
            "description": "Exact manufacturer, e.g. 'HPL'."
          },
          {
            "name": "install_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "installed | faulty | decommissioned",
              "title": "Install Status"
            },
            "description": "installed | faulty | decommissioned"
          },
          {
            "name": "phase_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "single | three",
              "title": "Phase Type"
            },
            "description": "single | three"
          },
          {
            "name": "install_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "whole_current | ct_operated",
              "title": "Install Type"
            },
            "description": "whole_current | ct_operated"
          },
          {
            "name": "build",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "legacy | v2",
              "title": "Build"
            },
            "description": "legacy | v2"
          },
          {
            "name": "dt_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Distribution transformer code, e.g. 'DT-001'.",
              "title": "Dt Code"
            },
            "description": "Distribution transformer code, e.g. 'DT-001'."
          },
          {
            "name": "hierarchy_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Any hierarchy code at any level, e.g. 'F-002' or 'Z-01'.",
              "title": "Hierarchy Code"
            },
            "description": "Any hierarchy code at any level, e.g. 'F-002' or 'Z-01'."
          },
          {
            "name": "has_location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only meters with/without coordinates.",
              "title": "Has Location"
            },
            "description": "Only meters with/without coordinates."
          },
          {
            "name": "with_issues",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only meters with/without data-quality flags.",
              "title": "With Issues"
            },
            "description": "Only meters with/without data-quality flags."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "One of: meter_id, serial_no, make, install_status, dt_code",
              "default": "meter_id",
              "title": "Sort"
            },
            "description": "One of: meter_id, serial_no, make, install_status, dt_code"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "asc",
              "title": "Order"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page (max 200).",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Items per page (max 200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MeterSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/meters/{meter_id}": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Get one meter",
        "description": "Full record including the normalised root-to-leaf hierarchy path and coordinates.\n\nUpstream this data has two incompatible encodings depending on the meter's build; both are normalised to this single shape.",
        "operationId": "get_meter_api_v1_meters__meter_id__get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Case-sensitive portal id, e.g. 'J100001'.",
              "title": "Meter Id"
            },
            "description": "Case-sensitive portal id, e.g. 'J100001'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meter"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such meter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hierarchy": {
      "get": {
        "tags": [
          "Hierarchy"
        ],
        "summary": "Get the network tree",
        "description": "The utility's network tree (Zone -> Circle -> Division -> Subdivision -> Substation -> Feeder -> DT), reconstructed by folding together every meter's root-to-leaf path. The portal only ever exposes one meter's path at a time.\n\n**Node identity is the full ancestor path, not the bare code.** Mid-level codes are not globally unique upstream - `D-01` occurs under three different circles - so `Z-01/C-01/D-01` and `Z-01/C-03/D-01` are distinct nodes. Collapsing them by code would silently merge unrelated parts of the network. See PROTOCOL.md.\n\n`meter_count` is cumulative: it counts meters anywhere beneath the node.",
        "operationId": "get_hierarchy_api_v1_hierarchy_get",
        "parameters": [
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 7,
              "minimum": 0,
              "description": "Levels to return below the root. 0 returns the root only.",
              "default": 7,
              "title": "Depth"
            },
            "description": "Levels to return below the root. 0 returns the root only."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HierarchyNode"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hierarchy/nodes/{node_id}/meters": {
      "get": {
        "tags": [
          "Hierarchy"
        ],
        "summary": "List meters under a subtree",
        "description": "Every meter whose path passes through the given node, paginated.",
        "operationId": "get_node_meters_api_v1_hierarchy_nodes__node_id__meters_get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Path id, e.g. 'Z-01/C-01'.",
              "title": "Node Id"
            },
            "description": "Path id, e.g. 'Z-01/C-01'."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page (max 200).",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Items per page (max 200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MeterSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No node with that path id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hierarchy/nodes/{node_id}": {
      "get": {
        "tags": [
          "Hierarchy"
        ],
        "summary": "Get a subtree",
        "description": "Fetch one node and its descendants by path id, e.g. `Z-01/C-01/D-01`. Use this to drill into a branch without transferring the whole tree.",
        "operationId": "get_hierarchy_node_api_v1_hierarchy_nodes__node_id__get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Path id, e.g. 'Z-01/C-01' or 'network'.",
              "title": "Node Id"
            },
            "description": "Path id, e.g. 'Z-01/C-01' or 'network'."
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 7,
              "minimum": 0,
              "default": 7,
              "title": "Depth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HierarchyNode"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No node with that path id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Envelope returned for **every** non-2xx response.\n\nDeclared so the OpenAPI document describes the real contract. Without it FastAPI\nadvertises its own ``HTTPValidationError`` for 422 and nothing at all for 404/5xx,\nwhich makes generated clients fail to deserialise every error they receive."
      },
...
      "HierarchyNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable path-based identity, e.g. 'Z-01/C-01/D-01'."
          },
          "level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyLevel"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null only for the synthetic 'network' root."
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count",
            "description": "Meters at or below this node."
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/HierarchyNode"
            },
            "type": "array",
            "title": "Children"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "meter_count"
        ],
        "title": "HierarchyNode",
        "description": "A node in the reconstructed network tree."
      },
...
      "InstallStatus": {
        "type": "string",
        "enum": [
          "installed",
          "faulty",
          "decommissioned"
        ],
        "title": "InstallStatus"
      },
...
      "Meter": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id",
            "description": "Portal identifier, e.g. 'J100001'. Case-sensitive."
          },
          "serial_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial No"
          },
          "make": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Make"
          },
          "phase_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PhaseType"
              },
              {
                "type": "null"
              }
            ]
          },
          "install_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstallStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "install_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstallType"
              },
              {
                "type": "null"
              }
            ]
          },
          "build": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeterBuild"
              },
              {
                "type": "null"
              }
            ]
          },
          "dt_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dt Code"
          },
          "hierarchy": {
            "items": {
              "$ref": "#/components/schemas/HierarchyRef"
            },
            "type": "array",
            "title": "Hierarchy",
            "description": "Root-to-leaf network path."
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoPoint"
              },
              {
                "type": "null"
              }
            ]
          },
          "data_quality": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Quality",
            "description": "Machine-readable slugs for anomalies detected in this record."
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "install_status",
          "install_type",
          "build",
          "dt_code"
        ],
        "title": "Meter",
        "description": "A smart meter, normalised across both upstream builds."
      },
...
      "MeterSummary": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "serial_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial No"
          },
          "make": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Make"
          },
          "phase_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PhaseType"
              },
              {
                "type": "null"
              }
            ]
          },
          "install_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstallStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "dt_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dt Code"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoPoint"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "install_status",
          "dt_code"
        ],
        "title": "MeterSummary",
        "description": "List-view projection. Omits the full hierarchy path to keep pages small."
      },
...
      "PageMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total_items",
          "total_pages",
          "has_next",
          "has_previous"
        ],
        "title": "PageMeta"
      },
      "Page_MeterSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeterSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[MeterSummary]"
      },
...
      "PhaseType": {
        "type": "string",
        "enum": [
          "single",
          "three"
        ],
        "title": "PhaseType"
      },
...
    }
  },
...
}